image-enhancer/image-enhancer/requirements.txt
flask>=2.3.0
pillow>=10.0.0
numpy>=1.24.0
